+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## [Unreleased]
 
+## [1.3.4] - 2022-07-18
+
 ### Improvements
 
 - add config option to make test plan optional (@orausch)
 - add comprehensive documentation (@oyamauchi)
+- add a `close` command (@joneshf)
+- allow `spr format` to be used without GitHub credentials
+- don't fail on requesting reviewers (@joneshf)
 
 ## [1.3.3] - 2022-06-27
 
@@ -120,3 +125,4 @@
 [1.3.1]: https://github.com/getcord/spr/releases/tag/v1.3.1
 [1.3.2]: https://github.com/getcord/spr/releases/tag/v1.3.2
 [1.3.3]: https://github.com/getcord/spr/releases/tag/v1.3.3
+[1.3.4]: https://github.com/getcord/spr/releases/tag/v1.3.4